refactor: move quest commands into cmd/camp/quest/ subpackage
Follows the same package-ownership pattern established by PR #170 for
dungeon, intent, leverage, etc. Quest command files move from flat
cmd/camp/quest_*.go into cmd/camp/quest/ with package quest. The root
command is exported as quest.Cmd and registered via a thin bridge file
(quest_register.go) in package main.
