+++ cmd/camp/quest/helpers.go
@@ -1,6 +1,6 @@
 //go:build dev
 
-package main
+package quest
 
 import (
 	"context"
